rtcrec: rename
